// enums/product.enum.ts
export enum ProductName {
    INTELLIJ = "Intellij",
    CLION = "CLion",
    RUSTROVER = "RustRover",
    ALL_PRODUCTS = "All",
    AI_PRO = "AiPro"
} 

export enum ProductSkuCode {
    INTELLIJ = "II",
    CLION = "CL",
    RUSTROVER = "RR",
    ALL_PRODUCTS = "ALL"
}

export enum ProductUrlCode {
    INTELLIJ = "idea",
    CLION = "clion",
    RUSTROVER = "rust",
} 